Update MKB2Controller.ahk -> 1.2.0
Added the ability to assign secondary keybinds for the same input. As a result the KeyList Helper interface is slightly different too.
Added the ability to use key combinations with LCtrl, LAlt, LShift.
Added the ability to not assign the mouse to any of the sticks.
Instead of using 2 methods for movement, now it'll always use the polling method, regardless of movement smoothing. Therefore, the checkbox for movement smoothing is also removed.
Changed the walk modifier assignment field into a button, to fix the funky behavior.
Fixed a bug with the walking speed slider.